+++ app/Http/Controllers/Usuarios/UsuariosController.php
@@ -78,22 +78,17 @@ class UsuariosController extends Controller
 
     DB::beginTransaction();
     try {
-      if ($actividad = Actividad::create($validated)) {
+      if ($usuario = Usuario::create($validator)) {
         DB::commit();
-        Alert::success('Acción completada', 'Actividad creada con éxito');
-        return redirect()->route('actividad.edit', $actividad);
+        Alert::success('Acción completada', 'Usuario creado con éxito');
+        return redirect()->route('usuario.modificar', $usuario->cedula);
       }
     } catch (Exception $error) {
       DB::rollBack();
       Log::error($error);
-      Alert::error('Oops!', 'Actividad no creada');
+      Alert::error('Oops!', 'Usuario no creado');
       return redirect()->back()->withInput();
     }
-
-    $usuario = Usuario::create($validator);
-
-    Alert::success('Acción completada', 'Usuario creado con éxito');
-    return redirect()->route('usuario.modificar', $usuario->cedula);
   }
 
   //ver modificar usuario
@@ -125,38 +120,33 @@ class UsuariosController extends Controller
 
     DB::beginTransaction();
     try {
-      if ($actividad = Actividad::create($validated)) {
+      if ($usuario->update($validator)) {
+        UsuarioProceso::where('usuario_id', Auth::id())->delete();
+        foreach ($validator['procesos'] ?? [] as $pro) {
+          $new = new UsuarioProceso;
+          $new->usuario_id = Auth::id();
+          $new->proceso_id = $pro;
+          $new->save();
+        }
+
+        UsuarioClientes::where('usuario_id', Auth::id())->delete();
+        foreach ($validator['clientes'] ?? [] as $cli) {
+          $new = new UsuarioClientes;
+          $new->usuario_id = Auth::id();
+          $new->cliente_id = $cli;
+          $new->save();
+        }
+
         DB::commit();
-        Alert::success('Acción completada', 'Actividad creada con éxito');
-        return redirect()->route('actividad.edit', $actividad);
+        Alert::success('Acción completada', 'Usuario modificado con éxito');
+        return redirect()->route('usuario.modificar', $usuario->cedula);
       }
     } catch (Exception $error) {
       DB::rollBack();
       Log::error($error);
-      Alert::error('Oops!', 'Actividad no creada');
+      Alert::error('Oops!', 'Usuario no modificado');
       return redirect()->back()->withInput();
     }
-
-    $usuario->update($validator);
-
-    UsuarioProceso::where('usuario_id', Auth::id())->delete();
-    foreach ($validator['procesos'] ?? [] as $pro) {
-      $new = new UsuarioProceso;
-      $new->usuario_id = Auth::id();
-      $new->proceso_id = $pro;
-      $new->save();
-    }
-
-    UsuarioClientes::where('usuario_id', Auth::id())->delete();
-    foreach ($validator['clientes'] ?? [] as $cli) {
-      $new = new UsuarioClientes;
-      $new->usuario_id = Auth::id();
-      $new->cliente_id = $cli;
-      $new->save();
-    }
-
-    Alert::success('Acción completada', 'Usuario modificado con éxito');
-    return redirect()->route('usuario.modificar', $usuario->cedula);
   }
